syncevo-dbus-server AutoTerm: fix signedness compiler warning
The compiler warned about arithmetic involving time_t and m_interval,
a guint. This patch solves that by making m_interval a time_t and
converting to it once from int.

Assigning a negative value to guint m_interval would have been strange,
and because time_t might be unsigned (?), better always set it to 0
in that case.