auto sync: user of autoSyncDelay/Interval must use seconds (fix for MB #10257)
The unit of SyncConfig::getAutoSyncDelay/Interval() had been changed to
seconds, but without adapting the syncevo-dbus-server code which uses
these values. Now it uses "seconds" whenever a delay or interval is
stored, like most of the rest of the code.

This patch also adds the unit to the comment for these variables.

While removing the factor of 60 in the conversion to the milliseconds
expected by the Timer class, I noticed that multiplying with 100
is not enough, it has to be 1000. Fixed.