Logging: changed importance of new SHOW level (fix for MB #5043)
The new SHOW level is not more severe than ERROR. It is more like
INFO, with the only difference being different formatting of the
output.

Making it more severe as in the original code had the effect that the
"level <= ERROR" check in SyncContext.cpp, which detects the first
error message, incorrectly recorded the first SHOW message as the
"first error encountered".

"level == ERROR" would have fixed this, but <= is better in case that
we later rearrange error levels and perhaps a "REALLY_FATAL_ERROR" (not
likely, but who knows...).