e_cal_check_timezones(): broken in EDS 2.30 due to change of error code (GNOME Bugzilla #611558, Moblin Bugzilla #9820)
The move to D-Bus changed the error code for "timezone cannot be
retrieved because it doesn't exist" from
E_CALENDAR_STATUS_OBJECT_NOT_FOUND to
E_CALENDAR_STATUS_INVALID_OBJECT. That latter error code was not
recognized by the code and caused e_cal_check_timezones() to fail with
that error ("Could not retrieve calendar time zone").

This patch fixes the problem by recognizing both the old and the new
error code.