D-Bus API: allow reading template for non-existent context (MB #10448)
A client which wants to create a new config in a non-default context should
always use
   config = GetConfig("<template>@context", true)
   <make changes>
   SetConfig(config)
to ensure that shared properties in that context are preserved.

A client shouldn't have to check whether that context exists. If it doesn't,
the server should simply ignore the context and use the default values from the
template.

This patch adds a test for this, based on the "type" property. The expectation
in the test that the "type" is preserved literally is wrong. "type" currently
must not be preserved, because different templates may have to set
different data formats (MB #10443).