packaging: fail harder if libnotify is missing (MB #1000).
When libnotify was not found, syncevo-dbus-server was silently
compiled without notifications, which happened in the MeeGo
snapshot .rpms due to missing build dependencies.

This patch introduces a --enable/disable-notify switch and aborts if
the feature is on (by default or explicitly requested) and the
development files are missing.

Also document this in README.packagers.