syncevo-dbus-server: redirect stdout (MB #5042)
Redirection of stdout was not activated and thus output of
synccompare was not sent via LogOutput messages.

Activating it shows that more work is necessary, although
it works in principle. Problems: line breaks are inserted
into long synccompare output and lines are swallowed.

Also, cout should not be used by the server itself. Instead
use the logging facilities.