LogRedirect with streams: better error handling (MB #5041)
When using streams, the caller of process() expects that once
the function returns, the streams were read until their end,
so that waitpid() will return eventually without deadlocking.

When an error occurs, this is not guaranteed, so we should
better return via an exception. Execute() was already
prepared for that.