command line + D-Bus: some parameters are already handled by parse() (MB #5043)
Things like "--sync-property ?" are handled by Cmdline::parse(), producing
output as the parameters are parsed. In this case, running the command line
again in the D-Bus server duplicates the output (once locally, once in the
server).

This patch adds a check for it, similar to the one in Cmdline::run() itself.