Cal + Mobical: apply the workaround only for Mobical (MB#10458)
Only apply conversion of alarm time to UTC time when server is
Mobical.net. The way is to define a global flag and set the flag
when 'remoterule' of 'Mobical' is matched.

The global flag is declared as a context variable and the script
reads it via built-in function 'SESSIONVAR'.