dbus api: change the signature of 'handler' of 'InfoRequest'
Currently the signature of argument 'handler' of 'InfoRequest'
is 's' on dbus server and sync ui. However, in the dbus api document,
its signature is still 'o'. Change it.

The reason for using a string and not an object path is that
when sending the initial request, the parameter has to be
empty. "" for a string is valid, but seems to be problematic
for an object path.