client-test: CLIENT_TEST_SIMPLE_UID
Our test data uses a UID with special characters, because that happens to
break "dumb" XML encoders/decoders and we want to find such issues. It also
breaks the EDS CalDAV backend. It can create such an event, but then cannot
access it  because the UID becomes of a URL used by WebDAV, and these special
characters are not escaped there (seen in evolution-data-server console output).

As a workaround, I simplified the test data to not use such a complicated UID
when CLIENT_TEST_SIMPLE_UID=1 is set. Still doesn't work, but the patch may
proof useful, so keep it.