vCard profile: disable Synthesis iPhone client TYPE extensions (MB #10462)
These extensions are useful when acting as server for a Synthesis iPhone
client, because then the server can preserve them.

But they are also confusing Evolution when vCards are imported with
these extensions are set. For example, an
ADR;TYPE=X-Synthesis-Ref0;TYPE=HOME is not shown as home address in
the GUI.

The correct solution would be to disable the extensions only when
writing into Evolution. We cannot do that due to technical limitations
("rule" parameter only works on properties).

So as a short term solution, choose the lesser evil and comment
out the extensions. Making Evolution work right is more important
than incremental enhancements of the server.