vCard profile: hide X-EVOLUTION-UI-SLOT in DevInf for Nokia N85 (MB #8868)
When we declare X-EVOLUTION-UI-SLOT as parameter in our DevInf, then
the N85 doesn't send telephone and email addresses to us. It seems to
get confused about what this unknown parameter is and then disables
the properties it is declared for.

As a workaround, do not "show" the parameter (per-parameter Synthesis
config option), but keep using it. The implication of that change is
minimal. The X-EVOLUTION-UI-SLOT value is used to preserve the order
of entries in the Evolution GUI. Declaring that our client supports
the value is only relevant when not sending it (because then a server
knows that it can drop its own value), but because we always send it,
we don't need to declare it.

In theory, a peer might avoid sending this value to us if we don't
declare it. In practice, peers are not that sophisticated, for good
reasons (like peers not declaring all of their properties). Synthesis
itself definitely still encodes the value even if not declared by
its peer.