syncevo-phone-config: fixed Python syntax errors
First, the "tempfile" variable clashed with the global tempfile module
and later was expected to be called "testfile". "tmpdir" is a better
description, because it is not the name of a file.

Second, the testConfig and testResult variables must explicitly
set as global variables. Variable assignments are local by default
in Python. Declaring the variables first is not strictly necessary,
but done to be consistent with other global variables.