syncevo-phone-config: avoid deprecated popen2, check for fatal status (MBC #1197)
Invoking "find" via popen2 is not particularly efficient. It also
triggers deprecation warnings. Better use native Python facilities,
like file pattern matching via glob. This is also shorter.

Instead of using an ad-hoc parser for the file, better use ConfigParser.
Some utility class is needed to massage our file content into shape
for it, because ConfigParser expects sections, which we don't write.

While we are at it, also check for premature exit of syncevolution.
The most likely reason for that is "killall -9 syncevolution", which
is currently the most reliable way to kill it because CTRL-C is not
always checked (needs to be fixed, though). Another reason are
fatal bugs in the binary, which should also abort testing.