syncevo-phone-config: avoid writing in user's .config while testing (MBC #1197)
Using a fixed, unique config name and a @test-phone context was meant
to avoid conflicts with user data, but that still has drawbacks:
- extra clean up code necessary
- cannot run multiple syncevo-phone-config calls in parallel

A better solution is to set XDG_CONFIG_HOME to the temporary directory.
That way we can leave the last test config in place, for review by the
user or a developer.

Another advantage is that it is safe to run without daemon mode, because
there really isn't any overlap left with the normal sync system. Not
using daemon mode
- has less overhead (no D-Bus involved)
- is easier to debug
- does not interfer with other, normal sync sessions

The patch renames the "syncevo" variable to ensure that all references
to it were updated. "syncevoCmd" is also a bit easier to search for.