syncevolution --version: missing infos about system shared libs
As reported on the mailing list, "syncevolution --version" does not
report about compatibility with shared system libraries (like "using
libecal-1.2.so.7") when using the daemon *and* no sync sessions has
been started.

The root cause is that the daemon does not initialize the system
compatibility when processing the "--version" command line. Once
something else does that, "--version" produces the correct output.

This is only a cosmetic issue, fixed by always initializing when
providing information.