syncevo-phone-config: avoid referencing invalid config (MBC #1197)
The code which set remoteIdentifier and SyncMLVersion in the template
copied from a "config" variable which was not set if a) not creating
a config earlier or b) the last source had no working config.

Case b actually occurred when running against a MeeGo netbook, which
has no "memo" database.

The fix is simple: only copy the values when at least one source has
them.