syncevo-phone-config: add "forced types" to tested configs (MBC #1197)
Types like "text/calendar" are implicitly turned into
"text/x-vcalendar" inside SyncEvolution when asking the phone to
sync. Forcing the type leads to a different SAN message and thus
is worth testing.

Forcing legacy types also leads to a slightly different sync session:
the SAN is the same as without forcing the type, but then we send
simpler CtCap information. This might make a difference when a phone
stumbles over too complex CtCap. But because such a phone should
work when disabling CtCap altogether, the forced legacy types are
not tested separately (would increase overall number of test cases).

Using --advanced with a --type for which no test data exists now fails
with an error instead of silently using empty items.