README: replaced with README.rst, added man page (MBC #690)
This patch adds the necessary autotools rules a) to use README.rst
as README in distribution archives and b) to build a syncevolution.1
man page out of it.

The later depends rst2man. --with-rst2man can be used to force man page
creation on or off. Without an explicit choice, it is only built when
the tool is found.

Version and compile date are inserted automatically.