README.html: built from README.rst with rst2html (MBC #690)
The main reason for bundling the .html version with the distribution
packages is that the syncevolution.org "Usage" page needs the
.html version. Building it as part of of the distribution packages
ensures that packages and documentation are in sync.

An additional benefit is that some users might prefer to read the
documentation in HTML.

As with man pages, building the HTML README depends on rst2html and
can be controlled via --with-rst2html.