Funambol, Memotoo: preserve meeting series when receiving update for detached recurrence (MBC #1916)
The RECURRENCE-ID is not preserved by these servers and also was not
preserved by SyncEvolution or Synthesis, because the servers do not
send CtCap information. As a result, updating a detached recurrence (=
a single instance in a meeting series) overwrote series in Evolution
and only left the updated recurrence.

This patch fixes that problem by copying the original RECURRENCE-ID
property from the original event which is to be updated. This is only
done when necessary, because it has to read the original event.