syncevo-dbus-server: first sync was done without libical time zone info (MBC #2435)
If (and only if) compiled with --enable-evolution-compatibility (as
the binaries on syncevolution.org), libical was only pulled into the
syncevo-dbus-server as part of running a sync. That was too late for
libsynthesis, which had already checked for libical earlier in that
sync session.

All following sync sessions then used libical time zone information.

The effect of not having libical time zone information were occasional
mismatches of time zones.