test-dbus.py: remove timeout before entering cleanup
When an exception occurred inside TestDBusSession.testSecondSession,
the timeout wasn't removed and would fire sometime later. If the
original problem was "server died", then the timeout code would
try to detach and cause another exception, which then stopped
the whole test-dbus.py run because it triggered outside of a test's
catch block.

This patch solves this by always removing the timeout in a "finally"
clause.