evo.supp: relaxed rule for icalparser_parse_string (MBC #1007)
Our valgrind suppression for a libical internal error did not
longer match because it included a long call stack. Simplified
the callstack to the one inside libical and the one level
inside SyncEvolution. This one location has been checked and
definitely is not our fault - other locations might be, so
don't suppress the error there.