Doc: enhanced D-Bus API documentation (MBC #1745)
Updated spec-to-docbook.xsl and the documentation in the .xml
files to improve formatting (added lists and paragraphs).

Added README explaining how this formatting works.

The updated spec-to-docbook.xsl no longer produces <reference>
elements in its output, which broke mergepages.xsl. Replaced
with some shell commands which concatenate the <refentry> elements
in a larger <reference> element with a title that also includes
the version.

Cross-references between the different APIs would be nice,
but don't work with the current spec-to-docbook.xsl (because
it only processes one API at a time?). Perhaps merge original
.xml files into one, then produce documentation? Nope, didn't
work. Seems that spec-to-docbook.xsl can only handle one interface
per file.